Pull util tests into new files and start refactoring tests
