+++ server/src/main/kotlin/services/utils/UpdateUtils.kt
@@ -1,20 +1,11 @@
-package services
+package services.utils
 
-import db.GameMutator
-import db.GamesQuery
-import db.WeeksQuery
-import dto.GameDTO
-import dto.WeekDTO
 import graphql.schema.DataFetchingEnvironment
 import graphql.schema.DataFetchingEnvironmentImpl
 import java.time.OffsetDateTime
 
 class UpdateUtils {
     companion object {
-
-
-
-
         fun gameStartedMoreThanXHoursAgo(gameTime: OffsetDateTime?, hoursAgo: Long): Boolean {
             return gameTime != null &&
                     OffsetDateTime.now().isAfter(gameTime.plusHours(hoursAgo))
