--- server/src/test/kotlin/ServiceRunnerTest.kt
@@ -1,1106 +1,0 @@
-import db.*
-import dto.*
-import graphql.schema.DataFetchingEnvironment
-import io.mockk.every
-import io.mockk.mockkClass
-import io.mockk.verify
-import org.junit.jupiter.api.Assertions.*
-import org.junit.jupiter.api.Test
-import services.*
-import java.io.FileNotFoundException
-import java.time.OffsetDateTime
-import java.util.*
-import kotlin.collections.HashMap
-
-class ServiceRunnerTest {
-
-    private val gameTimeInFarPast = OffsetDateTime.now().minusYears(1)
-
-    private val defaultGame: GameDTO = GameDTO("GB@CHI", "Week 7").apply {
-        gameTime = gameTimeInFarPast
-        result = "CHI"
-        id = UUID.randomUUID()
-        spread = -7.0
-    }
-
-    private val mockNflApi = mockkClass(NflApi::class)
-    private val mockMutator = mockkClass(GameMutator::class)
-
-    @Test
-    fun loadInfoForWeekMovesGameFromNflApiToDatabase() {
-        val gamesToDb = mutableListOf<GameDTO>()
-
-        val week = WeekDTO("Week 7")
-        val dbGame = defaultGame
-
-        every { mockNflApi.getWeek(any()) } returns listOf(dbGame)
-
-        every { mockMutator.putInDatabase(game = capture(gamesToDb)) } returns Unit
-
-        GameUpdateUtils.reloadGamesForWeek(week, mockNflApi, mockMutator)
-
-        verify(exactly = 1) { mockNflApi.getWeek(week) }
-        assertEquals(1, gamesToDb.size)
-        val gameToDb = gamesToDb[0]
-        assertEquals(dbGame.name, gameToDb.name)
-        assertEquals(dbGame.gameTime, gameToDb.gameTime)
-        assertEquals(dbGame.id, gameToDb.id)
-        assertEquals(dbGame.result, gameToDb.result)
-        assertEquals(dbGame.spread, gameToDb.spread)
-    }
-
-
-    @Test
-    fun loadInfoForWeekWithTwoGamesMovesBothFromNflApiToDatabase() {
-        val gamesToDb = mutableListOf<GameDTO>()
-
-        val week = WeekDTO("Week 7")
-        val dbGame1 = GameDTO("GB@CHI", "Week 7")
-        val dbGame2 = GameDTO("TB@NE", "Week 7").apply {
-            gameTime = gameTimeInFarPast.plusHours(1)
-            result = "TB"
-            id = UUID.randomUUID()
-            spread = 15.0
-        }
-
-        every { mockNflApi.getWeek(any()) } returns listOf(dbGame1, dbGame2)
-        every { mockMutator.putInDatabase(game = capture(gamesToDb)) } returns Unit
-
-        GameUpdateUtils.reloadGamesForWeek(week, mockNflApi, mockMutator)
-
-        verify(exactly = 1) { mockNflApi.getWeek(week) }
-        assertEquals(2, gamesToDb.size)
-        val secondGameToDb = gamesToDb[1]
-        assertEquals(dbGame2.name, secondGameToDb.name)
-        assertEquals(dbGame2.gameTime, secondGameToDb.gameTime)
-        assertEquals(dbGame2.id, secondGameToDb.id)
-        assertEquals(dbGame2.result, secondGameToDb.result)
-        assertEquals(dbGame2.spread, secondGameToDb.spread)
-    }
-
-    @Test
-    fun loadInfoForWeekAndFileNotFoundDoesNothing() {
-        every { mockNflApi.getWeek(any()) } throws FileNotFoundException("Mock failure")
-        every { mockMutator.putInDatabase(any()) } returns Unit
-
-        GameUpdateUtils.reloadGamesForWeek(WeekDTO("Week -1"), mockNflApi, mockMutator)
-
-        verify(exactly = 0) { mockMutator.putInDatabase(any()) }
-    }
-
-    @Test
-    fun updateGameDetailsForGameWithoutResultMovesGameFromNflToDb() {
-        val baseGame: GameDTO = GameDTO("GB@CHI", "Week 7").apply {
-            gameTime = gameTimeInFarPast
-            result = null
-            id = UUID.randomUUID()
-            spread = -7.0
-        }
-        every { mockNflApi.getGame(baseGame) } returns defaultGame
-        every { mockMutator.putInDatabase(any()) } returns Unit
-
-        GameUpdateUtils.updateDetailsForFinalGame(baseGame, mockNflApi, mockMutator)
-
-        verify(exactly = 1) { mockNflApi.getGame(baseGame) }
-        verify(exactly = 1) { mockMutator.putInDatabase(defaultGame) }
-    }
-
-    @Test
-    fun updateGameDetailsForGameWithoutIdDoesNothing() {
-        val baseGame: GameDTO = GameDTO("GB@CHI", "Week 7").apply {
-            gameTime = gameTimeInFarPast
-            result = null
-            id = null
-            spread = -7.0
-        }
-        every { mockNflApi.getGame(any()) } returns defaultGame
-        every { mockMutator.putInDatabase(any()) } returns Unit
-
-        GameUpdateUtils.updateDetailsForFinalGame(baseGame, mockNflApi, mockMutator)
-
-        verify(exactly = 0) { mockNflApi.getGame(any()) }
-        verify(exactly = 0) { mockMutator.putInDatabase(any()) }
-    }
-
-    @Test
-    fun updateGameDetailsForGameWithIdAndResultDoesNothing() {
-        val baseGame: GameDTO = GameDTO("GB@CHI", "Week 7").apply {
-            gameTime = gameTimeInFarPast
-            result = "CHI"
-            id = UUID.randomUUID()
-            spread = -7.0
-        }
-        every { mockNflApi.getGame(any()) } returns baseGame
-        every { mockMutator.putInDatabase(any()) } returns Unit
-
-        GameUpdateUtils.updateDetailsForFinalGame(baseGame, mockNflApi, mockMutator)
-
-        verify(exactly = 0) { mockNflApi.getGame(any()) }
-        verify(exactly = 0) { mockMutator.putInDatabase(any()) }
-    }
-
-    @Test
-    fun updateGameDetailsForGameWithoutResultAndAlmostTwoHoursAfterStartDoesNothing() {
-        val baseGame: GameDTO = GameDTO("GB@CHI", "Week 7").apply {
-            gameTime = OffsetDateTime.now().minusHours(2).plusMinutes(10)
-            result = null
-            id = UUID.randomUUID()
-            spread = -7.0
-        }
-        every { mockNflApi.getGame(any()) } returns baseGame
-        every { mockMutator.putInDatabase(any()) } returns Unit
-
-        GameUpdateUtils.updateDetailsForFinalGame(baseGame, mockNflApi, mockMutator)
-
-        verify(exactly = 0) { mockNflApi.getGame(any()) }
-        verify(exactly = 0) { mockMutator.putInDatabase(any()) }
-    }
-
-    @Test
-    fun updateGameDetailsForGameWithoutGameTimeDoesNothing() {
-        val baseGame: GameDTO = GameDTO("GB@CHI", "Week 7").apply {
-            gameTime = null
-            result = null
-            id = UUID.randomUUID()
-            spread = -7.0
-        }
-        every { mockNflApi.getGame(any()) } returns baseGame
-        every { mockMutator.putInDatabase(any()) } returns Unit
-
-        GameUpdateUtils.updateDetailsForFinalGame(baseGame, mockNflApi, mockMutator)
-
-        verify(exactly = 0) { mockNflApi.getGame(any()) }
-        verify(exactly = 0) { mockMutator.putInDatabase(any()) }
-    }
-
-    @Test
-    fun updateGameDetailsAndNflApiThrowsFileNotFoundSkipsPutInDatabase() {
-        val baseGame: GameDTO = GameDTO("GB@CHI", "Week 7").apply {
-            gameTime = gameTimeInFarPast
-            result = null
-            id = UUID.randomUUID()
-            spread = -7.0
-        }
-        every { mockNflApi.getGame(any()) } throws FileNotFoundException("Mock failure")
-
-        GameUpdateUtils.updateDetailsForFinalGame(baseGame, mockNflApi, mockMutator)
-
-        verify(exactly = 1) { mockNflApi.getGame(any()) }
-        verify(exactly = 0) { mockMutator.putInDatabase(any()) }
-    }
-
-    @Test
-    fun hasGamesMissingIdReturnsFalseWhenOnlyGameIsFinal() {
-        val mockGamesQuery = mockkClass(GamesQuery::class)
-        val mockWeeksQuery = mockkClass(WeeksQuery::class)
-        every { mockWeeksQuery.get() } returns listOf(WeekDTO("Week 1"))
-        every { mockGamesQuery.getGamesForWeek("Week 1") } returns listOf(
-            GameDTO("GB@CHI", "Week 1").apply {
-                result = "CHI"
-            }
-        )
-
-        val result = GameUpdateUtils.hasImmanentGamesMissingId(mockWeeksQuery, mockGamesQuery)
-
-        assertFalse(result)
-    }
-
-    @Test
-    fun hasGamesMissingIdReturnsTrueWhenFirstGameInSecondWeekHasNoId() {
-        val mockGamesQuery = mockkClass(GamesQuery::class)
-        val mockWeeksQuery = mockkClass(WeeksQuery::class)
-        every { mockWeeksQuery.get() } returns listOf(WeekDTO("Week 1"), WeekDTO("Week 2"))
-        every { mockGamesQuery.getGamesForWeek("Week 1") } returns listOf(
-            GameDTO("GB@CHI", "Week 1").apply {
-                result = "CHI"
-                id = UUID.randomUUID()
-            }
-        )
-        every { mockGamesQuery.getGamesForWeek("Week 2") } returns listOf(
-            GameDTO("DET@CHI", "Week 2").apply {
-                gameTime = OffsetDateTime.now().minusDays(1)
-            }
-        )
-
-        val result = GameUpdateUtils.hasImmanentGamesMissingId(mockWeeksQuery, mockGamesQuery)
-
-        assertTrue(result)
-    }
-
-    @Test
-    fun hasGamesMissingIdReturnsTrueWhenMissingIdAndTenMinutesBeforeGame() {
-        val mockGamesQuery = mockkClass(GamesQuery::class)
-        val mockWeeksQuery = mockkClass(WeeksQuery::class)
-        every { mockWeeksQuery.get() } returns listOf(WeekDTO("Week 1"))
-        every { mockGamesQuery.getGamesForWeek("Week 1") } returns listOf(
-            GameDTO("GB@CHI", "Week 1").apply {
-                gameTime = OffsetDateTime.now().plusMinutes(10)
-            }
-        )
-
-        val result = GameUpdateUtils.hasImmanentGamesMissingId(mockWeeksQuery, mockGamesQuery)
-
-        assertTrue(result)
-    }
-
-    @Test
-    fun hasGamesMissingIdResultAndGametimeReturnsFalse() {
-        val mockGamesQuery = mockkClass(GamesQuery::class)
-        val mockWeeksQuery = mockkClass(WeeksQuery::class)
-        every { mockWeeksQuery.get() } returns listOf(WeekDTO("Week 1"))
-        every { mockGamesQuery.getGamesForWeek("Week 1") } returns listOf(
-            GameDTO("GB@CHI", "Week 1")
-        )
-
-        val result = GameUpdateUtils.hasImmanentGamesMissingId(mockWeeksQuery, mockGamesQuery)
-
-        assertFalse(result)
-    }
-
-    @Test
-    fun hasGamesMissingIdReturnsFalseWhenMissingIdAndFifteenMinutesAfterGame() {
-        val mockGamesQuery = mockkClass(GamesQuery::class)
-        val mockWeeksQuery = mockkClass(WeeksQuery::class)
-        every { mockWeeksQuery.get() } returns listOf(WeekDTO("Week 1"))
-        every { mockGamesQuery.getGamesForWeek("Week 1") } returns listOf(
-            GameDTO("GB@CHI", "Week 1").apply {
-                gameTime = OffsetDateTime.now().plusMinutes(105)
-            }
-        )
-
-        val result = GameUpdateUtils.hasImmanentGamesMissingId(mockWeeksQuery, mockGamesQuery)
-
-        assertFalse(result)
-    }
-
-    @Test
-    fun hasGamesMissingIdReturnsTrueWhenSecondGameInSecondWeekHasNoId() {
-        val mockGamesQuery = mockkClass(GamesQuery::class)
-        val mockWeeksQuery = mockkClass(WeeksQuery::class)
-        every { mockWeeksQuery.get() } returns listOf(WeekDTO("Week 1"), WeekDTO("Week 2"))
-        every { mockGamesQuery.getGamesForWeek("Week 1") } returns listOf(
-            GameDTO("GB@CHI", "Week 1").apply {
-                result = "CHI"
-                gameTime = OffsetDateTime.now().minusDays(1)
-                id = UUID.randomUUID()
-            }
-        )
-        every { mockGamesQuery.getGamesForWeek("Week 2") } returns listOf(
-            GameDTO("DET@CHI", "Week 2").apply {
-                gameTime = OffsetDateTime.now().minusDays(1)
-                id = UUID.randomUUID()
-            },
-            GameDTO("NE@NYJ", "Week 2").apply {
-                gameTime = OffsetDateTime.now().minusDays(1)
-            }
-        )
-
-        val result = GameUpdateUtils.hasImmanentGamesMissingId(mockWeeksQuery, mockGamesQuery)
-
-        assertTrue(result)
-    }
-
-
-    @Test
-    fun randomNumberGeneratorGamesAreMakeForCurrentWeekWhenNotDoneAlready() {
-        val mockCurrentWeekQuery = mockkClass(CurrentWeekQuery::class)
-        val mockPicksQuery = mockkClass(UserPickQuery::class)
-        val mockGamesQuery = mockkClass(GamesQuery::class)
-        val mockPickMutator = mockkClass(UpdatePickMutator::class)
-        val mockRandomPickSelector = mockkClass(RandomPickSelector::class)
-
-        val mutatorEnvs = mutableListOf<DataFetchingEnvironment>()
-
-        val week = "Week 0"
-        val game = "DET@CHI"
-        val expectedPick = "CHI"
-        every { mockCurrentWeekQuery.getCurrentWeek() } returns WeekDTO(week)
-        every { mockGamesQuery.getGamesForWeek(week) } returns listOf(
-            GameDTO(game, week).apply {
-                gameTime = OffsetDateTime.now().plusDays(1)
-                id = UUID.randomUUID()
-            }
-        )
-        every { mockPicksQuery.getPicksForWeek(week) } returns listOf(
-            UserPicksDTO(UserDTO("RNG"))
-        )
-        every { mockPickMutator.get(capture(mutatorEnvs)) } returns true
-        every { mockRandomPickSelector.chooseRandomFor(game) } returns expectedPick
-
-        RngUpdateUtils.makeRngPicksForCurrentWeek(
-            mockCurrentWeekQuery,
-            mockGamesQuery,
-            mockPicksQuery,
-            mockPickMutator,
-            mockRandomPickSelector
-        )
-
-        assertEquals(1, mutatorEnvs.size)
-        val mutatorEnv = mutatorEnvs.first()
-        assertEquals("RNG", mutatorEnv.arguments["name"])
-        val userPick = mutatorEnv.arguments["userPick"] as HashMap<*, *>
-        assertEquals(week, userPick["week"])
-        assertEquals(game, userPick["game"])
-        assertEquals(expectedPick, userPick["pick"]!!)
-    }
-
-    @Test
-    fun randomNumberGeneratorDoesNotMakePicksForGamesBefore15MinutesOfGameStart() {
-        val mockCurrentWeekQuery = mockkClass(CurrentWeekQuery::class)
-        val mockPicksQuery = mockkClass(UserPickQuery::class)
-        val mockGamesQuery = mockkClass(GamesQuery::class)
-        val mockPickMutator = mockkClass(UpdatePickMutator::class)
-        val mockRandomPickSelector = mockkClass(RandomPickSelector::class)
-
-        val mutatorEnvs = mutableListOf<DataFetchingEnvironment>()
-
-        val week = "Week 0"
-        val game = "DET@CHI"
-        val expectedPick = "CHI"
-        every { mockCurrentWeekQuery.getCurrentWeek() } returns WeekDTO(week)
-        every { mockGamesQuery.getGamesForWeek(week) } returns listOf(
-            GameDTO(game, week).apply {
-                gameTime = OffsetDateTime.now().plusMinutes(14)
-                id = UUID.randomUUID()
-            }
-        )
-        every { mockPicksQuery.getPicksForWeek(week) } returns listOf(
-            UserPicksDTO(UserDTO("RNG"))
-        )
-        every { mockPickMutator.get(capture(mutatorEnvs)) } returns true
-        every { mockRandomPickSelector.chooseRandomFor(game) } returns expectedPick
-
-        RngUpdateUtils.makeRngPicksForCurrentWeek(
-            mockCurrentWeekQuery,
-            mockGamesQuery,
-            mockPicksQuery,
-            mockPickMutator,
-            mockRandomPickSelector
-        )
-
-        assertEquals(0, mutatorEnvs.size)
-    }
-
-    @Test
-    fun randomNumberGeneratorDoesNotMakePicksForGamesWithoutGametime() {
-        val mockCurrentWeekQuery = mockkClass(CurrentWeekQuery::class)
-        val mockPicksQuery = mockkClass(UserPickQuery::class)
-        val mockGamesQuery = mockkClass(GamesQuery::class)
-        val mockPickMutator = mockkClass(UpdatePickMutator::class)
-        val mockRandomPickSelector = mockkClass(RandomPickSelector::class)
-
-        val mutatorEnvs = mutableListOf<DataFetchingEnvironment>()
-
-        val week = "Week 0"
-        val game = "DET@CHI"
-        val expectedPick = "CHI"
-        every { mockCurrentWeekQuery.getCurrentWeek() } returns WeekDTO(week)
-        every { mockGamesQuery.getGamesForWeek(week) } returns listOf(
-            GameDTO(game, week).apply {
-                gameTime = null
-                id = UUID.randomUUID()
-            }
-        )
-        every { mockPicksQuery.getPicksForWeek(week) } returns listOf(
-            UserPicksDTO(UserDTO("RNG"))
-        )
-        every { mockPickMutator.get(capture(mutatorEnvs)) } returns true
-        every { mockRandomPickSelector.chooseRandomFor(game) } returns expectedPick
-
-        RngUpdateUtils.makeRngPicksForCurrentWeek(
-            mockCurrentWeekQuery,
-            mockGamesQuery,
-            mockPicksQuery,
-            mockPickMutator,
-            mockRandomPickSelector
-        )
-
-        assertEquals(0, mutatorEnvs.size)
-    }
-
-    @Test
-    fun randomNumberGeneratorPicksUseWeek() {
-        val mockCurrentWeekQuery = mockkClass(CurrentWeekQuery::class)
-        val mockPicksQuery = mockkClass(UserPickQuery::class)
-        val mockGamesQuery = mockkClass(GamesQuery::class)
-        val mockPickMutator = mockkClass(UpdatePickMutator::class)
-        val mockRandomPickSelector = mockkClass(RandomPickSelector::class)
-
-        val mutatorEnvs = mutableListOf<DataFetchingEnvironment>()
-
-        val week = "Week 1"
-        val game = "DET@CHI"
-        val expectedPick = "CHI"
-        every { mockCurrentWeekQuery.getCurrentWeek() } returns WeekDTO(week)
-        every { mockGamesQuery.getGamesForWeek(week) } returns listOf(
-            GameDTO(game, week).apply {
-                gameTime = OffsetDateTime.now().plusDays(1)
-                id = UUID.randomUUID()
-            }
-        )
-        every { mockPicksQuery.getPicksForWeek(week) } returns listOf(
-            UserPicksDTO(UserDTO("RNG"))
-        )
-        every { mockPickMutator.get(capture(mutatorEnvs)) } returns true
-        every { mockRandomPickSelector.chooseRandomFor(game) } returns expectedPick
-
-        RngUpdateUtils.makeRngPicksForCurrentWeek(
-            mockCurrentWeekQuery,
-            mockGamesQuery,
-            mockPicksQuery,
-            mockPickMutator,
-            mockRandomPickSelector
-        )
-
-        val mutatorEnv = mutatorEnvs.first()
-        val userPick = mutatorEnv.arguments["userPick"] as HashMap<*, *>
-        assertEquals(week, userPick["week"])
-    }
-
-    @Test
-    fun randomNumberGeneratorPicksUseGame() {
-        val mockCurrentWeekQuery = mockkClass(CurrentWeekQuery::class)
-        val mockPicksQuery = mockkClass(UserPickQuery::class)
-        val mockGamesQuery = mockkClass(GamesQuery::class)
-        val mockPickMutator = mockkClass(UpdatePickMutator::class)
-        val mockRandomPickSelector = mockkClass(RandomPickSelector::class)
-
-        val mutatorEnvs = mutableListOf<DataFetchingEnvironment>()
-
-        val week = "Week 1"
-        val game = "TB@NE"
-        val expectedPick = "CHI"
-        every { mockCurrentWeekQuery.getCurrentWeek() } returns WeekDTO(week)
-        every { mockGamesQuery.getGamesForWeek(week) } returns listOf(
-            GameDTO(game, week).apply {
-                gameTime = OffsetDateTime.now().plusDays(1)
-                id = UUID.randomUUID()
-            }
-        )
-        every { mockPicksQuery.getPicksForWeek(week) } returns listOf(
-            UserPicksDTO(UserDTO("RNG"))
-        )
-        every { mockPickMutator.get(capture(mutatorEnvs)) } returns true
-        every { mockRandomPickSelector.chooseRandomFor(game) } returns expectedPick
-
-        RngUpdateUtils.makeRngPicksForCurrentWeek(
-            mockCurrentWeekQuery,
-            mockGamesQuery,
-            mockPicksQuery,
-            mockPickMutator,
-            mockRandomPickSelector
-        )
-
-        val mutatorEnv = mutatorEnvs.first()
-        val userPick = mutatorEnv.arguments["userPick"] as HashMap<*, *>
-        assertEquals(game, userPick["game"])
-    }
-
-    @Test
-    fun randomNumberGeneratorPicksUseRandomNumberPickerForAwayTeam() {
-        val mockCurrentWeekQuery = mockkClass(CurrentWeekQuery::class)
-        val mockPicksQuery = mockkClass(UserPickQuery::class)
-        val mockGamesQuery = mockkClass(GamesQuery::class)
-        val mockPickMutator = mockkClass(UpdatePickMutator::class)
-        val mockRandomPickSelector = mockkClass(RandomPickSelector::class)
-
-        val mutatorEnvs = mutableListOf<DataFetchingEnvironment>()
-
-        val week = "Week 1"
-        val game = "TB@NE"
-        val expectedPick = "TB"
-        every { mockCurrentWeekQuery.getCurrentWeek() } returns WeekDTO(week)
-        every { mockGamesQuery.getGamesForWeek(week) } returns listOf(
-            GameDTO(game, week).apply {
-                gameTime = OffsetDateTime.now().plusDays(1)
-                id = UUID.randomUUID()
-            }
-        )
-        every { mockPicksQuery.getPicksForWeek(week) } returns listOf(
-            UserPicksDTO(UserDTO("RNG"))
-        )
-        every { mockPickMutator.get(capture(mutatorEnvs)) } returns true
-        every { mockRandomPickSelector.chooseRandomFor(game) } returns expectedPick
-
-        RngUpdateUtils.makeRngPicksForCurrentWeek(
-            mockCurrentWeekQuery,
-            mockGamesQuery,
-            mockPicksQuery,
-            mockPickMutator,
-            mockRandomPickSelector
-        )
-
-        val mutatorEnv = mutatorEnvs.first()
-        val userPick = mutatorEnv.arguments["userPick"] as HashMap<*, *>
-        assertEquals(expectedPick, userPick["pick"])
-    }
-
-    @Test
-    fun randomNumberGeneratorGamesAreMakeForCurrentWeekWithTwoGamesWhenNotDoneAlready() {
-        val mockCurrentWeekQuery = mockkClass(CurrentWeekQuery::class)
-        val mockPicksQuery = mockkClass(UserPickQuery::class)
-        val mockGamesQuery = mockkClass(GamesQuery::class)
-        val mockPickMutator = mockkClass(UpdatePickMutator::class)
-        val mockRandomPickSelector = mockkClass(RandomPickSelector::class)
-
-        val mutatorEnvs = mutableListOf<DataFetchingEnvironment>()
-
-        val week = "Week 0"
-        val game = "DET@CHI"
-        val expectedPick = "CHI"
-        every { mockCurrentWeekQuery.getCurrentWeek() } returns WeekDTO(week)
-        every { mockGamesQuery.getGamesForWeek(week) } returns listOf(
-            GameDTO("TB@NE", week).apply {
-                gameTime = OffsetDateTime.now().plusDays(1)
-                id = UUID.randomUUID()
-            },
-            GameDTO(game, week).apply {
-                gameTime = OffsetDateTime.now().plusDays(1)
-                id = UUID.randomUUID()
-            }
-        )
-        every { mockPicksQuery.getPicksForWeek(week) } returns listOf(
-            UserPicksDTO(UserDTO("RNG"))
-        )
-        every { mockPickMutator.get(capture(mutatorEnvs)) } returns true
-        every { mockRandomPickSelector.chooseRandomFor("TB@NE") } returns "TB"
-        every { mockRandomPickSelector.chooseRandomFor(game) } returns expectedPick
-
-        RngUpdateUtils.makeRngPicksForCurrentWeek(
-            mockCurrentWeekQuery,
-            mockGamesQuery,
-            mockPicksQuery,
-            mockPickMutator,
-            mockRandomPickSelector
-        )
-
-        assertEquals(2, mutatorEnvs.size)
-        val mutatorEnv = mutatorEnvs[1]
-        assertEquals("RNG", mutatorEnv.arguments["name"])
-        val userPick = mutatorEnv.arguments["userPick"] as HashMap<*, *>
-        assertEquals(week, userPick["week"])
-        assertEquals(game, userPick["game"])
-        assertEquals(expectedPick, userPick["pick"]!!)
-    }
-
-
-    @Test
-    fun randomNumberGeneratorSkipsPicksThatAreAlreadyMade() {
-        val mockCurrentWeekQuery = mockkClass(CurrentWeekQuery::class)
-        val mockPicksQuery = mockkClass(UserPickQuery::class)
-        val mockGamesQuery = mockkClass(GamesQuery::class)
-        val mockPickMutator = mockkClass(UpdatePickMutator::class)
-        val mockRandomPickSelector = mockkClass(RandomPickSelector::class)
-
-        val mutatorEnvs = mutableListOf<DataFetchingEnvironment>()
-
-        val week = "Week 0"
-        val game = "DET@CHI"
-        val expectedPick = "CHI"
-        every { mockCurrentWeekQuery.getCurrentWeek() } returns WeekDTO(week)
-        val pickedGame = "TB@NE"
-        every { mockGamesQuery.getGamesForWeek(week) } returns listOf(
-            GameDTO(game, week).apply {
-                gameTime = OffsetDateTime.now().plusDays(1)
-                id = UUID.randomUUID()
-            },
-            GameDTO(pickedGame, week).apply {
-                gameTime = OffsetDateTime.now().plusDays(1)
-                id = UUID.randomUUID()
-            }
-        )
-        every { mockPicksQuery.getPicksForWeek(week) } returns listOf(
-            UserPicksDTO(UserDTO("RNG")).apply {
-                picks = arrayListOf(PickDTO(pickedGame, "NE"))
-            }
-        )
-        every { mockPickMutator.get(capture(mutatorEnvs)) } returns true
-        every { mockRandomPickSelector.chooseRandomFor(pickedGame) } returns "TB"
-        every { mockRandomPickSelector.chooseRandomFor(game) } returns expectedPick
-
-        RngUpdateUtils.makeRngPicksForCurrentWeek(
-            mockCurrentWeekQuery,
-            mockGamesQuery,
-            mockPicksQuery,
-            mockPickMutator,
-            mockRandomPickSelector
-        )
-
-        assertEquals(1, mutatorEnvs.size)
-        val mutatorEnv = mutatorEnvs.first()
-        val userPick = mutatorEnv.arguments["userPick"] as HashMap<*, *>
-        assertEquals(game, userPick["game"])
-    }
-
-    @Test
-    fun randomNumberGeneratorUsesOnlyRngPicks() {
-        val mockCurrentWeekQuery = mockkClass(CurrentWeekQuery::class)
-        val mockPicksQuery = mockkClass(UserPickQuery::class)
-        val mockGamesQuery = mockkClass(GamesQuery::class)
-        val mockPickMutator = mockkClass(UpdatePickMutator::class)
-        val mockRandomPickSelector = mockkClass(RandomPickSelector::class)
-
-        val mutatorEnvs = mutableListOf<DataFetchingEnvironment>()
-
-        val week = "Week 0"
-        val game = "DET@CHI"
-        val expectedPick = "CHI"
-        every { mockCurrentWeekQuery.getCurrentWeek() } returns WeekDTO(week)
-        val pickedGame = "TB@NE"
-        every { mockGamesQuery.getGamesForWeek(week) } returns listOf(
-            GameDTO(game, week).apply {
-                gameTime = OffsetDateTime.now().plusDays(1)
-                id = UUID.randomUUID()
-            },
-            GameDTO(pickedGame, week).apply {
-                gameTime = OffsetDateTime.now().plusDays(1)
-                id = UUID.randomUUID()
-            }
-        )
-        every { mockPicksQuery.getPicksForWeek(week) } returns listOf(
-            UserPicksDTO(UserDTO("Someone Else")).apply {
-                picks = arrayListOf(PickDTO(pickedGame, "NE"))
-            },
-            UserPicksDTO(UserDTO("RNG"))
-        )
-        every { mockPickMutator.get(capture(mutatorEnvs)) } returns true
-        every { mockRandomPickSelector.chooseRandomFor(pickedGame) } returns "TB"
-        every { mockRandomPickSelector.chooseRandomFor(game) } returns expectedPick
-
-        RngUpdateUtils.makeRngPicksForCurrentWeek(
-            mockCurrentWeekQuery,
-            mockGamesQuery,
-            mockPicksQuery,
-            mockPickMutator,
-            mockRandomPickSelector
-        )
-
-        assertEquals(2, mutatorEnvs.size)
-    }
-
-    @Test
-    fun oneVegasPickUpdatesPickForCurrentWeek() {
-        val mockCurrentWeekQuery = mockkClass(CurrentWeekQuery::class)
-        val mockGamesQuery = mockkClass(GamesQuery::class)
-        val mockPickMutator = mockkClass(UpdatePickMutator::class)
-        val mockGameMutator = mockkClass(GameMutator::class)
-        val mockVegasPicksApi = mockkClass(VegasPicksApi::class)
-
-        val pickMutatorEnvs = mutableListOf<DataFetchingEnvironment>()
-        val gameMutations = mutableListOf<GameDTO>()
-
-        val week = "Week 0"
-        val game = "DET@CHI"
-        val expectedPick = "CHI"
-        every { mockCurrentWeekQuery.getCurrentWeek() } returns WeekDTO(week)
-        every { mockGamesQuery.getGamesForWeek(week) } returns listOf(
-            GameDTO(game, week).apply {
-                gameTime = OffsetDateTime.now().plusDays(1)
-                id = UUID.randomUUID()
-            }
-        )
-
-        every { mockPickMutator.get(capture(pickMutatorEnvs)) } returns true
-        every { mockGameMutator.putInDatabase(capture(gameMutations)) } returns Unit
-        every { mockVegasPicksApi.getVegasPicks() } returns listOf(
-            PickWithSpreadDTO(game, expectedPick, -7.0)
-        )
-
-        VegasUpdateUtils.updateVegasPicks(
-            mockCurrentWeekQuery,
-            mockGamesQuery,
-            mockGameMutator,
-            mockPickMutator,
-            mockVegasPicksApi
-        )
-
-        assertEquals(1, pickMutatorEnvs.size)
-        val mutatorEnv = pickMutatorEnvs.first()
-        assertEquals("Vegas", mutatorEnv.arguments["name"])
-        val userPick = mutatorEnv.arguments["userPick"] as HashMap<*, *>
-        assertEquals(week, userPick["week"])
-        assertEquals(game, userPick["game"])
-        assertEquals(expectedPick, userPick["pick"]!!)
-    }
-
-    @Test
-    fun oneVegasPickUpdatesWithCurrentWeek() {
-        val mockCurrentWeekQuery = mockkClass(CurrentWeekQuery::class)
-        val mockGamesQuery = mockkClass(GamesQuery::class)
-        val mockPickMutator = mockkClass(UpdatePickMutator::class)
-        val mockGameMutator = mockkClass(GameMutator::class)
-        val mockVegasPicksApi = mockkClass(VegasPicksApi::class)
-
-        val pickMutatorEnvs = mutableListOf<DataFetchingEnvironment>()
-        val gameMutations = mutableListOf<GameDTO>()
-
-        val week = "Week 1"
-        val game = "DET@CHI"
-        val expectedPick = "CHI"
-        every { mockCurrentWeekQuery.getCurrentWeek() } returns WeekDTO(week)
-        every { mockGamesQuery.getGamesForWeek(week) } returns listOf(
-            GameDTO(game, week).apply {
-                gameTime = OffsetDateTime.now().plusDays(1)
-                id = UUID.randomUUID()
-            }
-        )
-
-        every { mockPickMutator.get(capture(pickMutatorEnvs)) } returns true
-        every { mockGameMutator.putInDatabase(capture(gameMutations)) } returns Unit
-        every { mockVegasPicksApi.getVegasPicks() } returns listOf(
-            PickWithSpreadDTO(game, expectedPick, -7.0)
-        )
-
-        VegasUpdateUtils.updateVegasPicks(
-            mockCurrentWeekQuery,
-            mockGamesQuery,
-            mockGameMutator,
-            mockPickMutator,
-            mockVegasPicksApi
-        )
-
-
-        val mutatorEnv = pickMutatorEnvs.first()
-        val userPick = mutatorEnv.arguments["userPick"] as HashMap<*, *>
-        assertEquals(week, userPick["week"])
-    }
-
-    @Test
-    fun oneVegasPickUpdatesPickWithGame() {
-        val mockCurrentWeekQuery = mockkClass(CurrentWeekQuery::class)
-        val mockGamesQuery = mockkClass(GamesQuery::class)
-        val mockPickMutator = mockkClass(UpdatePickMutator::class)
-        val mockGameMutator = mockkClass(GameMutator::class)
-        val mockVegasPicksApi = mockkClass(VegasPicksApi::class)
-
-        val pickMutatorEnvs = mutableListOf<DataFetchingEnvironment>()
-        val gameMutations = mutableListOf<GameDTO>()
-
-        val week = "Week 1"
-        val game = "TB@NE"
-        val expectedPick = "TB"
-        every { mockCurrentWeekQuery.getCurrentWeek() } returns WeekDTO(week)
-        every { mockGamesQuery.getGamesForWeek(week) } returns listOf(
-            GameDTO(game, week).apply {
-                gameTime = OffsetDateTime.now().plusDays(1)
-                id = UUID.randomUUID()
-            }
-        )
-
-        every { mockPickMutator.get(capture(pickMutatorEnvs)) } returns true
-        every { mockGameMutator.putInDatabase(capture(gameMutations)) } returns Unit
-        every { mockVegasPicksApi.getVegasPicks() } returns listOf(
-            PickWithSpreadDTO(game, expectedPick, -7.0)
-        )
-
-        VegasUpdateUtils.updateVegasPicks(
-            mockCurrentWeekQuery,
-            mockGamesQuery,
-            mockGameMutator,
-            mockPickMutator,
-            mockVegasPicksApi
-        )
-
-
-        val mutatorEnv = pickMutatorEnvs.first()
-        val userPick = mutatorEnv.arguments["userPick"] as HashMap<*, *>
-        assertEquals(game, userPick["game"])
-    }
-
-    @Test
-    fun oneVegasPickUpdatesPickWithPick() {
-        val mockCurrentWeekQuery = mockkClass(CurrentWeekQuery::class)
-        val mockGamesQuery = mockkClass(GamesQuery::class)
-        val mockPickMutator = mockkClass(UpdatePickMutator::class)
-        val mockGameMutator = mockkClass(GameMutator::class)
-        val mockVegasPicksApi = mockkClass(VegasPicksApi::class)
-
-        val pickMutatorEnvs = mutableListOf<DataFetchingEnvironment>()
-        val gameMutations = mutableListOf<GameDTO>()
-
-        val week = "Week 1"
-        val game = "TB@NE"
-        val expectedPick = "TB"
-        every { mockCurrentWeekQuery.getCurrentWeek() } returns WeekDTO(week)
-        every { mockGamesQuery.getGamesForWeek(week) } returns listOf(
-            GameDTO(game, week).apply {
-                gameTime = OffsetDateTime.now().plusDays(1)
-                id = UUID.randomUUID()
-            }
-        )
-
-        every { mockPickMutator.get(capture(pickMutatorEnvs)) } returns true
-        every { mockGameMutator.putInDatabase(capture(gameMutations)) } returns Unit
-        every { mockVegasPicksApi.getVegasPicks() } returns listOf(
-            PickWithSpreadDTO(game, expectedPick, -7.0)
-        )
-
-        VegasUpdateUtils.updateVegasPicks(
-            mockCurrentWeekQuery,
-            mockGamesQuery,
-            mockGameMutator,
-            mockPickMutator,
-            mockVegasPicksApi
-        )
-
-
-        val mutatorEnv = pickMutatorEnvs.first()
-        val userPick = mutatorEnv.arguments["userPick"] as HashMap<*, *>
-        assertEquals(expectedPick, userPick["pick"])
-    }
-
-    @Test
-    fun oneVegasPickUpdatesSpreadForGame() {
-        val mockCurrentWeekQuery = mockkClass(CurrentWeekQuery::class)
-        val mockGamesQuery = mockkClass(GamesQuery::class)
-        val mockPickMutator = mockkClass(UpdatePickMutator::class)
-        val mockGameMutator = mockkClass(GameMutator::class)
-        val mockVegasPicksApi = mockkClass(VegasPicksApi::class)
-
-        val pickMutatorEnvs = mutableListOf<DataFetchingEnvironment>()
-        val gameMutations = mutableListOf<GameDTO>()
-
-        val week = "Week 1"
-        val game = "TB@NE"
-        val expectedPick = "TB"
-        every { mockCurrentWeekQuery.getCurrentWeek() } returns WeekDTO(week)
-        val expectedGameTime = OffsetDateTime.now().plusDays(1)
-        val randomUUID = UUID.randomUUID()
-        every { mockGamesQuery.getGamesForWeek(week) } returns listOf(
-            GameDTO(game, week).apply {
-                gameTime = expectedGameTime
-                id = randomUUID
-            }
-        )
-
-        every { mockPickMutator.get(capture(pickMutatorEnvs)) } returns true
-        every { mockGameMutator.putInDatabase(capture(gameMutations)) } returns Unit
-        val expectedSpread = -7.0
-        every { mockVegasPicksApi.getVegasPicks() } returns listOf(
-            PickWithSpreadDTO(game, expectedPick, expectedSpread)
-        )
-
-        VegasUpdateUtils.updateVegasPicks(
-            mockCurrentWeekQuery,
-            mockGamesQuery,
-            mockGameMutator,
-            mockPickMutator,
-            mockVegasPicksApi
-        )
-
-
-        assertEquals(1, gameMutations.size)
-        val gameMutation = gameMutations.first()
-        assertEquals(game, gameMutation.name)
-        assertEquals(week, gameMutation.week)
-        assertEquals(expectedSpread, gameMutation.spread)
-        assertEquals(randomUUID, gameMutation.id)
-    }
-
-    @Test
-    fun oneVegasPickUpdatesSpreadForVaryingSpread() {
-        val mockCurrentWeekQuery = mockkClass(CurrentWeekQuery::class)
-        val mockGamesQuery = mockkClass(GamesQuery::class)
-        val mockPickMutator = mockkClass(UpdatePickMutator::class)
-        val mockGameMutator = mockkClass(GameMutator::class)
-        val mockVegasPicksApi = mockkClass(VegasPicksApi::class)
-
-        val pickMutatorEnvs = mutableListOf<DataFetchingEnvironment>()
-        val gameMutations = mutableListOf<GameDTO>()
-
-        val week = "Week 1"
-        val game = "TB@NE"
-        val expectedPick = "TB"
-        every { mockCurrentWeekQuery.getCurrentWeek() } returns WeekDTO(week)
-        val expectedGameTime = OffsetDateTime.now().plusDays(1)
-        val randomUUID = UUID.randomUUID()
-        every { mockGamesQuery.getGamesForWeek(week) } returns listOf(
-            GameDTO(game, week).apply {
-                gameTime = expectedGameTime
-                id = randomUUID
-            }
-        )
-
-        every { mockPickMutator.get(capture(pickMutatorEnvs)) } returns true
-        every { mockGameMutator.putInDatabase(capture(gameMutations)) } returns Unit
-        val expectedSpread = -87.0
-        every { mockVegasPicksApi.getVegasPicks() } returns listOf(
-            PickWithSpreadDTO(game, expectedPick, expectedSpread)
-        )
-
-        VegasUpdateUtils.updateVegasPicks(
-            mockCurrentWeekQuery,
-            mockGamesQuery,
-            mockGameMutator,
-            mockPickMutator,
-            mockVegasPicksApi
-        )
-
-
-        assertEquals(1, gameMutations.size)
-        val gameMutation = gameMutations.first()
-        assertEquals(expectedSpread, gameMutation.spread)
-    }
-
-    @Test
-    fun twoVegasPicksUpdatesPick() {
-        val mockCurrentWeekQuery = mockkClass(CurrentWeekQuery::class)
-        val mockGamesQuery = mockkClass(GamesQuery::class)
-        val mockPickMutator = mockkClass(UpdatePickMutator::class)
-        val mockGameMutator = mockkClass(GameMutator::class)
-        val mockVegasPicksApi = mockkClass(VegasPicksApi::class)
-
-        val pickMutatorEnvs = mutableListOf<DataFetchingEnvironment>()
-        val gameMutations = mutableListOf<GameDTO>()
-
-        val week = "Week 0"
-        val game = "DET@CHI"
-        val expectedPick = "CHI"
-        every { mockCurrentWeekQuery.getCurrentWeek() } returns WeekDTO(week)
-        every { mockGamesQuery.getGamesForWeek(week) } returns listOf(
-            GameDTO("TB@NE", week).apply {
-                gameTime = OffsetDateTime.now().plusDays(1)
-                id = UUID.randomUUID()
-            },
-            GameDTO(game, week).apply {
-                gameTime = OffsetDateTime.now().plusDays(1)
-                id = UUID.randomUUID()
-            }
-        )
-
-        every { mockPickMutator.get(capture(pickMutatorEnvs)) } returns true
-        every { mockGameMutator.putInDatabase(capture(gameMutations)) } returns Unit
-        every { mockVegasPicksApi.getVegasPicks() } returns listOf(
-            PickWithSpreadDTO("TB@NE", "TB", -7.0),
-            PickWithSpreadDTO(game, expectedPick, 3.0)
-        )
-
-        VegasUpdateUtils.updateVegasPicks(
-            mockCurrentWeekQuery,
-            mockGamesQuery,
-            mockGameMutator,
-            mockPickMutator,
-            mockVegasPicksApi
-        )
-
-        assertEquals(2, pickMutatorEnvs.size)
-        val mutatorEnv = pickMutatorEnvs[1]
-        val userPick = mutatorEnv.arguments["userPick"] as HashMap<*, *>
-        assertEquals(game, userPick["game"])
-        assertEquals(expectedPick, userPick["pick"]!!)
-    }
-
-    @Test
-    fun vegasPickWithoutMatchingGameInDbIsNotMade() {
-        val mockCurrentWeekQuery = mockkClass(CurrentWeekQuery::class)
-        val mockGamesQuery = mockkClass(GamesQuery::class)
-        val mockPickMutator = mockkClass(UpdatePickMutator::class)
-        val mockGameMutator = mockkClass(GameMutator::class)
-        val mockVegasPicksApi = mockkClass(VegasPicksApi::class)
-
-        val pickMutatorEnvs = mutableListOf<DataFetchingEnvironment>()
-        val gameMutations = mutableListOf<GameDTO>()
-
-        val week = "Week 0"
-        val game = "DET@CHI"
-        val expectedPick = "CHI"
-        every { mockCurrentWeekQuery.getCurrentWeek() } returns WeekDTO(week)
-        every { mockGamesQuery.getGamesForWeek(week) } returns listOf()
-
-        every { mockPickMutator.get(capture(pickMutatorEnvs)) } returns true
-        every { mockGameMutator.putInDatabase(capture(gameMutations)) } returns Unit
-        every { mockVegasPicksApi.getVegasPicks() } returns listOf(
-            PickWithSpreadDTO(game, expectedPick, 3.0)
-        )
-
-        VegasUpdateUtils.updateVegasPicks(
-            mockCurrentWeekQuery,
-            mockGamesQuery,
-            mockGameMutator,
-            mockPickMutator,
-            mockVegasPicksApi
-        )
-
-        assertEquals(0, pickMutatorEnvs.size)
-    }
-
-    @Test
-    fun vegasPicksForGamesAboutToStartAreNotUpdated() {
-        val mockCurrentWeekQuery = mockkClass(CurrentWeekQuery::class)
-        val mockGamesQuery = mockkClass(GamesQuery::class)
-        val mockPickMutator = mockkClass(UpdatePickMutator::class)
-        val mockGameMutator = mockkClass(GameMutator::class)
-        val mockVegasPicksApi = mockkClass(VegasPicksApi::class)
-
-        val pickMutatorEnvs = mutableListOf<DataFetchingEnvironment>()
-        val gameMutations = mutableListOf<GameDTO>()
-
-        val week = "Week 0"
-        val game = "DET@CHI"
-        val expectedPick = "CHI"
-        every { mockCurrentWeekQuery.getCurrentWeek() } returns WeekDTO(week)
-        every { mockGamesQuery.getGamesForWeek(week) } returns listOf(
-            GameDTO(game, week).apply {
-                gameTime = OffsetDateTime.now().plusMinutes(14)
-                id = UUID.randomUUID()
-            }
-        )
-
-        every { mockPickMutator.get(capture(pickMutatorEnvs)) } returns true
-        every { mockGameMutator.putInDatabase(capture(gameMutations)) } returns Unit
-        every { mockVegasPicksApi.getVegasPicks() } returns listOf(
-            PickWithSpreadDTO("TB@NE", "TB", -7.0),
-            PickWithSpreadDTO(game, expectedPick, 3.0)
-        )
-
-        VegasUpdateUtils.updateVegasPicks(
-            mockCurrentWeekQuery,
-            mockGamesQuery,
-            mockGameMutator,
-            mockPickMutator,
-            mockVegasPicksApi
-        )
-
-        assertEquals(0, pickMutatorEnvs.size)
-    }
-
-    @Test
-    fun vegasPicksForGamesWithoutGametimeAreNotUpdated() {
-        val mockCurrentWeekQuery = mockkClass(CurrentWeekQuery::class)
-        val mockGamesQuery = mockkClass(GamesQuery::class)
-        val mockPickMutator = mockkClass(UpdatePickMutator::class)
-        val mockGameMutator = mockkClass(GameMutator::class)
-        val mockVegasPicksApi = mockkClass(VegasPicksApi::class)
-
-        val pickMutatorEnvs = mutableListOf<DataFetchingEnvironment>()
-        val gameMutations = mutableListOf<GameDTO>()
-
-        val week = "Week 0"
-        val game = "DET@CHI"
-        val expectedPick = "CHI"
-        every { mockCurrentWeekQuery.getCurrentWeek() } returns WeekDTO(week)
-        every { mockGamesQuery.getGamesForWeek(week) } returns listOf(
-            GameDTO(game, week).apply {
-                gameTime = null
-                id = UUID.randomUUID()
-            }
-        )
-
-        every { mockPickMutator.get(capture(pickMutatorEnvs)) } returns true
-        every { mockGameMutator.putInDatabase(capture(gameMutations)) } returns Unit
-        every { mockVegasPicksApi.getVegasPicks() } returns listOf(
-            PickWithSpreadDTO("TB@NE", "TB", -7.0),
-            PickWithSpreadDTO(game, expectedPick, 3.0)
-        )
-
-        VegasUpdateUtils.updateVegasPicks(
-            mockCurrentWeekQuery,
-            mockGamesQuery,
-            mockGameMutator,
-            mockPickMutator,
-            mockVegasPicksApi
-        )
-
-        assertEquals(0, pickMutatorEnvs.size)
-    }
-}
